Refactor the way we send packets and remove ugly hack
We now store the plaintext data in the send array and encrypt it on each send attempt.
This prevents us from sending packets from our send array that were encrypted with
an old shared key that the recipient no longer has the means to decrypt. This in
turn allows us to remove the ugly hack where we were keeping he last shared key
around to decrypt recent messages