No longer send chat ID hash in plaintext packet headers
This was used to identify the group that a groupchat packet was intended for
however due to the public nature of DHT groupchats, it would expose the fact
that someone was in a particular group to anyone who looked.

We now use a hash of the recipient's group permanent public key for handshake
packets, and a hash of the recipient's session public key associated
with the sender's connection for all other packets.

This significantly reduces the ability of adversaries from determining that a
target is in a particular group.