Refactor plaintext header/group identification (again)
We now ditch using hashes in the plaintext header and just use the full
permanent public key of the peer that the packet is intended for. The
reason for this change is because using a public key hash in the header
would slightly weaken the cryptography due to the fact that you wouldn't
be able to have multiple keys in multiple groups that create the same
hash.

This is part 1 of a 2-part refactor. To be continued...