Improve peer loading from saved lists
When we load a peerlist on client start or from the timeout
list we no longer send oob TCP requests unless necessary,
and we send invite requests instead of peer info requests.

This fixes an issue with the timeout list not working and
should make connecting to peers after starting a client
more reliable.