Make some improvements to error handling inbound packets
 A packet handler should return success if we want to
send the sender an ack even if the packet contained
invalid data. This is to prevent cases where an out
of sync packet that contains invalid data gets stuck
in the send queue and blocks subsequent packets.

We only want to return an error if we want the sender
to know about the failure.