Add peer roles checksum to sync packets and fix sync issues
We now send a checksum comprised of all our confirmed peers' role
plus the first byte of their respective public key in order to
ensure that all peers see the same set of roles for their peer
list. We also detect when we have a conflict in our list and reset
our shared state version to zero in order to force a sync request.