Never stop trying to handshake with saved peers
When a group is private we were only sending a few handshake
packets before giving up forever. Since there was no DHT telling
us about other peers in the group, this means that if we failed
to handshake on our first attempt, we would never be able to
connect until a manual reconnect attempt.