Add DHT queries to private API
This commit adds functionality for clients to send DHT getnodes
requests to their DHT peers. It also allows clients to make
queries on DHT nodes that they receive in a getnodes response,
including port, public key, and IP address (as a null terminated
string).

A typical use case would be for a DHT network crawler, but it
could also be used for understanding and diagnosing DHT network
behaviour.