Refactor DHT API
We get rid of the opaque Tox_Dht_Node struct and pass public key,
ip and port directly to and from functions. This removes the
need for node query functions and allows us to send arbitrary
get nodes requests instead of requiring an object passed to
us from the callback.