New feature: ability for founder to set voice state
There are three voice states. If set to ALL, everyone may speak as normal
with the exception of observers. If set to MODERATOR, only moderators
and the founder may speak. If set to FOUNDER, only the founder may speak.

This state has no effect on topic setting or the ability to send private
messages.