Refactor onion_client.c do_friends() to reduce network traffic
Onion announce requests tend to be very spammy and excessive. We don't
really need to be searching the DHT for offline friends 10+ times per
minute. This function alone accounted for nearly half of all tox's
network traffic.

With this commit we now do friend announces on a 2 minute interval,
except when a friend has been added, where they will be on a
temporary interval of 3 seconds. This allows clients to aggressively
search the DHT when a new contact is added or on initial boot,
and then to periodically query the DHT for offline friends without
needlessly spamming it thereafter.

My initial tests using the netprof branch show more than a 50%
reduction in tox-related network traffic with UDP enabled, and
a ~33% reduction with UDP disabled. However, these changes would
have a network-wide impact (fewer inbound queries), so those
numbers are probably an underestimate.