Improve how we establish direct connections with friends
We now send our IP_Port to friends in ping packets when we establish
new connections. This forces us to keep trying to establish a UDP
connection if we initially connect via TCP. This addresses an issue
where we often make an initial connection via TCP and have to wait
a long time before a UDP connection gets established.

We only make 10 attempts before giving up so that we aren't wasting too
much bandwidth trying to establish direct connections with friends
who have UDP disabled, or who we otherwise won't be able to establish
a direct connection with due to holepunching failure.