Improve how we establish direct connections with friends
We now send our IP_Port to friends in ping packets when we establish
new connections. This forces us to keep trying to establish a UDP
connection if we initially connect via TCP. This addresses an issue
where we often make an initial connection via TCP and have to wait
a long time before a UDP connection gets established.

We make 5 attempts approximately every 3 minute interval.