refactor: Make prune_gc_sanctions_list more obviously correct.
`target_ext_pk` is uninitialised before the loop, and is definitely
initialised after the loop *iff* `sanction` isn't NULL. If it is NULL,
then `target_ext_pk` is not read. This is verifiable, but not obvious.
PVS Studio isn't smart enough to figure that out, so it complains. I
consider it a valid complaint, and the new code is clearer: we now call
the processing part of the function directly with initialised values
instead of relying on that initialisation coinciding with `sanction`s
nullity.