feat: Build the fun utils when cross-compiling to Windows
While fun utils are of very low quality, both code-wise and usage-wise
-- not checking for failed mallocs, not offering usage instructions,
etc., there are a couple of them that Windows users might find useful,
like the vanity key generators or a savedata creator, for example.

The building of the fun utils was broken on Windows due to the utils
failing to find sodium.h, as no libsodium include dirs were set on the
fun utils.