fix(autotools): add tox_log_level.h to public headers list
The tox_log_level.h header file was added as a public API header in
version 0.2.21, but was not included in the libtoxcore_la_include_HEADERS
variable in toxcore/Makefile.inc. This caused the header to not be
installed when building with autotools, leading to compilation errors
for applications that include tox_options.h (which depends on
tox_log_level.h, like qTox).

This patch adds tox_log_level.h to the list of installed headers,
matching the behavior of the CMake build system.

Commit: c8344726378e3be2277e82734ffd39c40c15d2e4
Fixes: compilation error "tox_log_level.h: No such file or directory"